+++ geofileops/geoops.py
@@ -339,7 +339,7 @@ def apply(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -352,7 +352,8 @@ def apply(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(f"Start apply on {input_path}")
+    logger = logging.getLogger("geofileops.apply")
+    logger.info(f"Start on {input_path}")
 
     return _geoops_gpd.apply(
         input_path=Path(input_path),
@@ -449,7 +450,7 @@ def buffer(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -539,8 +540,9 @@ def buffer(
     .. |buffer_mitre_10| image:: ../_static/images/buffer_mitre_10.png
         :alt: Buffer with mitre=1.0
     """  # noqa: E501
+    logger = logging.getLogger("geofileops.buffer")
     logger.info(
-        f"Start buffer on {input_path} "
+        f"Start, on {input_path} "
         f"(distance: {distance}, quadrantsegments: {quadrantsegments})"
     )
 
@@ -624,6 +626,8 @@ def clip_by_geometry(
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
     """
+    logger = logging.getLogger("geofileops.clip_by_geometry")
+    logger.info(f"Start, on {input_path}")
     return _geoops_ogr.clip_by_geometry(
         input_path=Path(input_path),
         output_path=Path(output_path),
@@ -680,7 +684,7 @@ def convexhull(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -693,7 +697,8 @@ def convexhull(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(f"Start convexhull on {input_path}")
+    logger = logging.getLogger("geofileops.convexhull")
+    logger.info(f"Start, on {input_path}")
 
     return _geoops_sql.convexhull(
         input_path=Path(input_path),
@@ -754,7 +759,8 @@ def delete_duplicate_geometries(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(f"Start delete_duplicate_geometries on {input_path}")
+    logger = logging.getLogger("geofileops.delete_duplicate_geometries")
+    logger.info(f"Start, on {input_path}")
 
     return _geoops_sql.delete_duplicate_geometries(
         input_path=Path(input_path),
@@ -916,7 +922,7 @@ def dissolve(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -942,7 +948,8 @@ def dissolve(
             # If an empty list of geometry columns is passed, convert it to None
             groupby_columns = None
 
-    logger.info(f"Start dissolve on {input_path} to {output_path}")
+    logger = logging.getLogger("geofileops.dissolve")
+    logger.info(f"Start, on {input_path} to {output_path}")
     return _geoops_gpd.dissolve(
         input_path=Path(input_path),
         output_path=Path(output_path),
@@ -994,6 +1001,8 @@ def export_by_bounds(
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
     """
+    logger = logging.getLogger("geofileops.export_by_bounds")
+    logger.info(f"Start, on {input_path}")
     return _geoops_ogr.export_by_bounds(
         input_path=Path(input_path),
         output_path=Path(output_path),
@@ -1047,7 +1056,7 @@ def isvalid(
         validate_attribute_data (bool, optional): True to validate if all attribute data
             can be read. Defaults to False.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -1068,7 +1077,8 @@ def isvalid(
         )
 
     # Go!
-    logger.info(f"Start isvalid on {input_path}")
+    logger = logging.getLogger("geofileops.isvalid")
+    logger.info(f"Start, on {input_path}")
     return _geoops_sql.isvalid(
         input_path=Path(input_path),
         output_path=output_path,
@@ -1141,7 +1151,7 @@ def makevalid(
             can be read. Raises an exception if an error is found, as this type of error
             cannot be fixed using makevalid. Defaults to False.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -1154,7 +1164,8 @@ def makevalid(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(f"Start makevalid on {input_path}")
+    logger = logging.getLogger("geofileops.makevalid")
+    logger.info(f"Start, on {input_path}")
 
     if gridsize is None:
         gridsize = 0.0
@@ -1233,7 +1244,8 @@ def warp(
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
     """
-    logger.info(f"Start warp on {input_path}")
+    logger = logging.getLogger("geofileops.warp")
+    logger.info(f"Start, on {input_path}")
     _geoops_ogr.warp(
         input_path=Path(input_path),
         output_path=Path(output_path),
@@ -1371,7 +1383,8 @@ def select(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(f"Start select on {input_path}")
+    logger = logging.getLogger("geofileops.select")
+    logger.info(f"Start, on {input_path}")
 
     # Convert force_output_geometrytype to GeometryType (if necessary)
     if force_output_geometrytype is not None:
@@ -1453,7 +1466,7 @@ def simplify(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -1466,7 +1479,8 @@ def simplify(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(f"Start simplify on {input_path} with tolerance {tolerance}")
+    logger = logging.getLogger("geofileops.simplify")
+    logger.info(f"Start, on {input_path} with tolerance {tolerance}")
     if isinstance(algorithm, str):
         algorithm = SimplifyAlgorithm(algorithm)
 
@@ -1480,6 +1494,7 @@ def simplify(
             columns=columns,
             explodecollections=explodecollections,
             gridsize=gridsize,
+            keep_empty_geoms=keep_empty_geoms,
             where_post=where_post,
             nb_parallel=nb_parallel,
             batchsize=batchsize,
@@ -1497,6 +1512,7 @@ def simplify(
             columns=columns,
             explodecollections=explodecollections,
             gridsize=gridsize,
+            keep_empty_geoms=keep_empty_geoms,
             where_post=where_post,
             nb_parallel=nb_parallel,
             batchsize=batchsize,
@@ -1504,9 +1520,9 @@ def simplify(
         )
 
 
-################################################################################
+# ------------------------
 # Operations on two layers
-################################################################################
+# ------------------------
 
 
 def clip(
@@ -1571,7 +1587,7 @@ def clip(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -1588,7 +1604,8 @@ def clip(
     .. |clip_result| image:: ../_static/images/clip_result.png
         :alt: Clip result
     """  # noqa: E501
-    logger.info(f"Start clip on {input_path} with {clip_path} to {output_path}")
+    logger = logging.getLogger("geofileops.clip")
+    logger.info(f"Start on {input_path} with {clip_path} to {output_path}")
     return _geoops_sql.clip(
         input_path=Path(input_path),
         clip_path=Path(clip_path),
@@ -1656,7 +1673,7 @@ def erase(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -1674,7 +1691,8 @@ def erase(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(f"Start erase on {input_path} with {erase_path} to {output_path}")
+    logger = logging.getLogger("geofileops.erase")
+    logger.info(f"Start, on {input_path} with {erase_path} to {output_path}")
     return _geoops_sql.erase(
         input_path=Path(input_path),
         erase_path=Path(erase_path),
@@ -1744,7 +1762,7 @@ def export_by_location(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -1757,8 +1775,9 @@ def export_by_location(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
+    logger = logging.getLogger("geofileops.export_by_location")
     logger.info(
-        f"Start export_by_location: select from {input_to_select_from_path} "
+        f"export_by_location: select from {input_to_select_from_path} "
         f"interacting with {input_to_compare_with_path} to {output_path}"
     )
     return _geoops_sql.export_by_location(
@@ -1822,7 +1841,7 @@ def export_by_distance(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -1835,8 +1854,9 @@ def export_by_distance(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
+    logger = logging.getLogger("geofileops.export_by_distance")
     logger.info(
-        f"Start export_by_distance: select from {input_to_select_from_path} within "
+        f"select from {input_to_select_from_path} within "
         f"max_distance of {max_distance} from {input_to_compare_with_path} "
         f"to {output_path}"
     )
@@ -1913,7 +1933,7 @@ def identity(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -1931,9 +1951,8 @@ def identity(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(
-        f"Start identity between {input1_path} and {input2_path} to {output_path}"
-    )
+    logger = logging.getLogger("geofileops.identity")
+    logger.info(f"Start, between {input1_path} and {input2_path} to {output_path}")
     return _geoops_sql.identity(
         input1_path=Path(input1_path),
         input2_path=Path(input2_path),
@@ -1983,9 +2002,8 @@ def split(
         FutureWarning,
         stacklevel=2,
     )
-    logger.info(
-        f"Start identity between {input1_path} and {input2_path} to {output_path}"
-    )
+    logger = logging.getLogger("geofileops.identity")
+    logger.info(f"Start,  between {input1_path} and {input2_path} to {output_path}")
     return _geoops_sql.identity(
         input1_path=Path(input1_path),
         input2_path=Path(input2_path),
@@ -2107,7 +2125,7 @@ def intersection(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -2120,9 +2138,8 @@ def intersection(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(
-        f"Start intersection between {input1_path} and {input2_path} to {output_path}"
-    )
+    logger = logging.getLogger("geofileops.intersection")
+    logger.info(f"Start, between {input1_path} and {input2_path} to {output_path}")
     return _geoops_sql.intersection(
         input1_path=Path(input1_path),
         input2_path=Path(input2_path),
@@ -2228,7 +2245,7 @@ def join_by_location(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -2241,10 +2258,8 @@ def join_by_location(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger.info(
-        f"Start join_by_location: select from {input1_path} joined with "
-        f"{input2_path} to {output_path}"
-    )
+    logger = logging.getLogger("geofileops.join_by_location")
+    logger.info(f"select from {input1_path} joined with {input2_path} to {output_path}")
     return _geoops_sql.join_by_location(
         input1_path=Path(input1_path),
         input2_path=Path(input2_path),
@@ -2313,7 +2328,7 @@ def join_nearest(
         output_layer (str, optional): output layer name. If None, the output_path stem
             is used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -2321,10 +2336,8 @@ def join_nearest(
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
     """
-    logger.info(
-        f"Start join_nearest: select from {input1_path} joined with "
-        f"{input2_path} to {output_path}"
-    )
+    logger = logging.getLogger("geofileops.join_nearest")
+    logger.info(f"select from {input1_path} joined with {input2_path} to {output_path}")
     return _geoops_sql.join_nearest(
         input1_path=Path(input1_path),
         input2_path=Path(input2_path),
@@ -2481,7 +2494,7 @@ def select_two_layers(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -2530,10 +2543,8 @@ def select_two_layers(
         <a href="https://github.com/geofileops/geofileops/blob/main/geofileops/util/geofileops_sql.py" target="_blank">geofileops_sql.py</a>
 
     """  # noqa: E501
-    logger.info(
-        f"Start select_two_layers: select from {input1_path} and {input2_path} "
-        f"to {output_path}"
-    )
+    logger = logging.getLogger("geofileops.select_two_layers")
+    logger.info(f"select from {input1_path} and {input2_path} to {output_path}")
     return _geoops_sql.select_two_layers(
         input1_path=Path(input1_path),
         input2_path=Path(input2_path),
@@ -2613,7 +2624,7 @@ def symmetric_difference(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduc
@@ -2632,9 +2643,10 @@ def symmetric_difference(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
+    logger = logging.getLogger("geofileops.symmetric_difference")
     logger.info(
-        f"Start symmetric_difference of {input1_path} and {input2_path} "
-        f"to {output_path}"
+        f"Start, with input1: {input1_path}, "
+        f"input2 {input2_path}, output: {output_path}"
     )
     return _geoops_sql.symmetric_difference(
         input1_path=Path(input1_path),
@@ -2713,7 +2725,7 @@ def union(
             including e.g. explodecollections. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
-            Defaults to -1: use all available processors.
+            Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
             batch. A smaller batch size, possibly in combination with a
             smaller nb_parallel, will reduce the memory usage.
@@ -2731,8 +2743,10 @@ def union(
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
+    logger = logging.getLogger("geofileops.union")
     logger.info(
-        f"Start union: select from {input1_path} and {input2_path} to {output_path}"
+        f"Start, with input1: {input1_path}, input2: {input2_path}, output: "
+        f"{output_path}"
     )
     return _geoops_sql.union(
         input1_path=Path(input1_path),